feat: implement quantile-based slippage protection with P95/P99 statistical learning
- Add quantile_impact.rs module with historical fill observation tracking
- Implement per-pool learning with size-aware categorization (Small/Medium/Large)
- Add percentile-based min_out calculation (P95/P99 configurable)
- Integrate quantile calculator into SniperEngine with compute_min_out() helper
- Record fill observations in transaction reconciliation
- Replace hardcoded Orca slippage calculations with quantile-based approach
- Add config options with adaptive slippage fallback
- Fix Solana SDK 3.x compilation errors (OptionSerializer handling)
- Add comprehensive documentation in QUANTILE_SLIPPAGE.md
- Update config.example.toml with quantile configuration section
- Mark TASKS.md line 93 as DONE with complete implementation details

+++ src/engine/mod.rs
@@ -638,7 +638,7 @@ impl Engine {
                                     skipped_events = skipped,
                                     "arbitrage_task: buffer lagged (normal during high activity)"
                                 );
-                                events_in_batch += skipped;
+                                events_in_batch += skipped as usize;
                                 break; // Scan immediately after lag
                             }
                             None => {
